requirements.txt
confusable-homoglyphs==3.1.1
dj-database-url==0.5.0
Django==1.11
django-bootstrap3==10.0.1
django-heroku==0.3.1
django-registration==2.4.1
flake8==3.5.0
gunicorn==19.8.1
jedi==0.12.0
mccabe==0.6.1
parso==0.2.0
Pillow==5.1.0
psycopg2==2.7.4
pycodestyle==2.3.1
pyflakes==1.6.0
python-dotenv==0.8.2
pytz==2018.4
whitenoise==3.3.1
